third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/object-tag-javascript-url-expected.txt

CONSOLE ERROR: line 8: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?wait-for-load=1&q=%3Cobject%20id=%27object%27%20data=%27javascript:alert(document.domain)%27%3E%3C/object%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
